rename: 상수 값 디렉토리 이름 수정
constants로 오타 수정
